Moves the code for actually creating the simulation from xml into the simulation class and distributes the creation of objects from xml into each relivant object.
