+++ simulation/__init__.py
@@ -1,3 +1,4 @@
+from simulation.custom_xml import local_test_xml_location, xml_bool, createPygameVector2, createPygameVector3, createPygameColor, load_XML, load_XML_without_schema
 from simulation.commands import commands as command_options
 from simulation.layer import Layer, Layer_2D, Layer_3D, Layer_Mixed
 from simulation.main import Simulation, RenderMode
