+++ simulation/custom_xml/__init__.py
@@ -1,0 +1,1 @@
+from simulation.custom_xml.xml_loading import local_schema_location, local_test_xml_location, xml_bool, createPygameVector2, createPygameVector3, createPygameColor, load_XML, load_XML_without_schema
